Move projections.py to skycrs.py and update references.
